Solving 3 Sudoku Boards
For an OOP assignment I was tasked with implementing the use of the enumeration function and File manipulation.

In the version of the Project I changed the core definition of the Board Class. Due to the fact that I needed to be able to define each square class as the subsequent sudoku board.